refactor(eventstream): revert module name to singular transformStreamController
